Update equity table styling
